+++ tracker/src/main/java/com/jay/oss/tracker/registry/StorageNodeRegistry.java
@@ -40,7 +40,7 @@ public class StorageNodeRegistry {
      * @return {@link List<StorageNodeInfo>} storage节点列表
      * @throws Exception e
      */
-    public List<StorageNodeInfo> selectUploadNode(long size, int replica) throws Exception {
+    public List<StorageNodeInfo> selectUploadNode(int size, int replica) throws Exception {
         List<StorageNodeInfo> aliveNodes = registry.aliveNodes();
         return replicaSelector.select(aliveNodes, size, replica);
     }
